flesh out netlink implementation
A lot of this is boiler plate code copied from the other drivers, that
may get consolidated later.  But for now, start fleshing out the L2
netlink implementation.  In particular, add the Tx/Rx paths for L2
Ethernet traffic.  This requires updates to the drivers/netlink package
for latest Netlinker definitions.