call notifyDown() directly, under d.hw lock
If code becomes re-entrant, for example, where notify callback for link
DOWN in stack trys to send a pkt thru driver, but sending a pkt in
driver acquires d.hw lock, so we deadlock.  Seems unlikely stack would
want to send after link is DOWN, not sure?  We can revisit if
deadlocking is a problem.